+++ jest.config.js
@@ -4,11 +4,5 @@ module.exports = {
   collectCoverageFrom: [
     'lib/**/*.js',
     '!lib/plugin.js'
-  ],
-  moduleNameMapper: {
-    '^~/(.*)$': '<rootDir>/lib/$1',
-    '^~~$': '<rootDir>',
-    '^@@$': '<rootDir>',
-    '^@/(.*)$': '<rootDir>/lib/$1'
-  }
+  ]
 }