+++ CHANGELOG.md
@@ -2,6 +2,103 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.0.0-beta.8](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.7...v1.0.0-beta.8) (2019-07-22)
+
+
+
+## [1.0.0-beta.7](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.6...v1.0.0-beta.7) (2019-07-18)
+
+
+
+## [1.0.0-beta.6](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.5...v1.0.0-beta.6) (2019-07-11)
+
+
+
+## [1.0.0-beta.5](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.4...v1.0.0-beta.5) (2019-07-03)
+
+
+
+## [1.0.0-beta.4](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.3...v1.0.0-beta.4) (2019-06-26)
+
+
+
+## [1.0.0-beta.3](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.2...v1.0.0-beta.3) (2019-06-19)
+
+
+
+## [1.0.0-beta.2](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2019-06-11)
+
+
+
+## [1.0.0-beta.1](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-beta.0...v1.0.0-beta.1) (2019-06-06)
+
+
+
+## [1.0.0-beta.0](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.6...v1.0.0-beta.0) (2019-06-04)
+
+
+### Bug Fixes
+
+* apply vuetify vars only for sass files ([#25](https://github.com/nuxt-community/vuetify-module/issues/25)) ([f8f0392](https://github.com/nuxt-community/vuetify-module/commit/f8f0392))
+* don't override sassLoader data ([3c1c1e6](https://github.com/nuxt-community/vuetify-module/commit/3c1c1e6))
+* fix tooltip activator behavior ([feeacd8](https://github.com/nuxt-community/vuetify-module/commit/feeacd8))
+* use theme primary color for toolbar ([3bc02cc](https://github.com/nuxt-community/vuetify-module/commit/3bc02cc))
+
+
+### Features
+
+* customVariables ([d2b3b8a](https://github.com/nuxt-community/vuetify-module/commit/d2b3b8a))
+* roboto font display swap ([feac192](https://github.com/nuxt-community/vuetify-module/commit/feac192))
+
+
+
+# [1.0.0-alpha.6](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.5...v1.0.0-alpha.6) (2019-05-18)
+
+
+
+# [1.0.0-alpha.5](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.4...v1.0.0-alpha.5) (2019-05-02)
+
+
+
+# [1.0.0-alpha.4](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.3...v1.0.0-alpha.4) (2019-04-18)
+
+
+### Bug Fixes
+
+* always setup sass loader options ([1709f2b](https://github.com/nuxt-community/vuetify-module/commit/1709f2b))
+
+
+
+# [1.0.0-alpha.3](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.2...v1.0.0-alpha.3) (2019-04-18)
+
+
+### Bug Fixes
+
+* hotfix when module options is not set in config ([4cfa0ab](https://github.com/nuxt-community/vuetify-module/commit/4cfa0ab))
+
+
+
+# [1.0.0-alpha.2](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.1...v1.0.0-alpha.2) (2019-04-18)
+
+
+
+# [1.0.0-alpha.1](https://github.com/nuxt-community/vuetify-module/compare/v1.0.0-alpha.0...v1.0.0-alpha.1) (2019-04-17)
+
+
+
+# [1.0.0-alpha.0](https://github.com/nuxt-community/vuetify-module/compare/v0.5.5...v1.0.0-alpha.0) (2019-04-10)
+
+
+### Features
+
+* upgrade to new vuetify 2 spec ([6ff1910](https://github.com/nuxt-community/vuetify-module/commit/6ff1910))
+
+
+### BREAKING CHANGES
+
+* Removed `css`, implemented `assets` option
+
+
 ### [0.5.7](https://github.com/nuxt-community/vuetify-module/compare/v0.5.6...v0.5.7) (2019-07-22)